requirements.txt
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==44.0.3
exceptiongroup==1.3.0
fastapi==0.117.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
linkify-it-py==2.0.3
loguru==0.7.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.4
nonebot-adapter-console==0.9.0
nonebot-adapter-onebot==2.4.6
nonebot-adapter-qq==1.6.5
nonebot-adapter-telegram==0.1.0b20
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-localstore==0.7.4
nonebot2==2.4.3
nonechat==0.6.1
nonestorage==0.1.0
platformdirs==4.4.0
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pygtrie==2.5.0
python-dotenv==1.1.1
PyYAML==6.0.2
rich==14.1.0
sniffio==1.3.1
starlette==0.48.0
textual==3.7.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
uc-micro-py==1.0.3
uvicorn==0.36.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
